prettier the added code
